Refactor mock factories to include optional overrides for path resolver and data fetcher services. Enhance schema validator and data fetcher implementations for improved test flexibility. Update related tests to utilize new mock structures.
